+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGE LOG
 
 * Improved the error info class
 * Return empty body and no content type if we can't match a displayer
+* Conformed to json api standards
 
 
 ## V2.0.1 (26/05/2015)